+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added path tracing support for directional, point and spot lights, as well as emission from Lit and Unlit.
 - Added non temporal version of SSAO.
 - Added more detailed ray tracing stats in the debug window
+- Added Disc area light (bake only)
 
 ### Fixed
 - Sorting, undo, labels, layout in the Lighting Explorer.
@@ -123,6 +124,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Improved quality of SSAO.
 - Moved RequestShadowMapRendering() back to public API.
 - Update HDRP DXR Wizard with an option to automatically clone the hdrp config package and setup raytracing to 1 in shaders file.
+- Simplified Light's type API regrouping the logic in one place (Check type in HDAdditionalLightData)
 
 ## [7.1.1] - 2019-09-05